backend/requirements.txt
click==8.0.3
colorama==0.4.4
Flask==2.0.3
Flask-Bcrypt
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
importlib-metadata==4.11.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
marshmallow==3.15.0
packaging==21.3
PyJWT
PyMySQL==1.0.2
pyparsing==3.0.7
python-dateutil
six==1.16.0
SQLAlchemy==1.4.31
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
